backend/requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation and Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
